// todo-webapp/src/generated/todo-api.ts
export interface components {
    schemas: {
        Todo: {
            /** @description todo identifier */
            id: string;
            /** @description what needs to be done */
            text: string;
            /** @description whether the todo is done */
            completed: boolean;
            /**
             * Format: date-time
             * @description creation timestamp
             */
            createdAt: string;
        };
        NewTodo: {
            /** @description what needs to be done */
            text: string;
        };
        TodoList: {
            /** @description total matching todos */
            count: number;
            /** @description relative URI of the next page */
            next?: string | null;
            /** @description relative URI of the previous page */
            previous?: string | null;
            data: components["schemas"]["Todo"][];
        };
        Error: {
            /** @description HTTP or application error code */
            code: number;
            /** @description short human-readable label */
            message: string;
            /** @description detailed explanation */
            description?: string;
            /** @description URI to documentation */
            moreInfo?: string;
        };
    };
    responses: never;
    parameters: {
        /** @description caller identity injected by the gateway from the validated token */
        UserId: string;
        /** @description todo identifier */
        TodoId: string;
    };
    requestBodies: never;
    headers: never;
    pathItems: never;
}